+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-api-helpers` will be documented in this file.
 
+## 0.0.3 - 2021-10-23
+
+- Add a phone number converter for changing from WhatsApp format to WHMCS format
+- Add tests for phone number conversion
+
 ## 0.0.2 - 2021-10-16
 
 - Add generic CURL DELETE